packages/w3up-python-client/requirements.txt
requests
python-dotenv
ipfs-car